pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
Flask = "^2.0.3"
youtube_dl = "^2021.12.17"
ffmpeg = "^1.4"
PyNaCl = "^1.5.0"
discord = "^1.7.3"
discord-pretty-help = "^1.3.3"
quart = "^0.16.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"